+++ packages/runtime/package.json
@@ -10,7 +10,8 @@
     "./client": "./esm/index.js",
     "./server": "./esm/index.server.js",
     "./jsx-runtime": "./esm/jsx-runtime.js",
-    "./jsx-dev-runtime": "./esm/jsx-dev-runtime.js"
+    "./jsx-dev-runtime": "./esm/jsx-dev-runtime.js",
+    "./matchRoutes": "./esm/matchRoutes.js"
   },
   "files": [
     "esm",
